Create step 2
The speech recognition function is integrated with the Ohbot. 

Ohbot now listens and speaks. But with some disabilities. 

The threading function is used but needs to be studied deeply and edited for more clarification. 

**The trigger function to start a conversation must be defined. 
For now, the program just starts when executed. This must be changed to execute when a "key word / trigger" word is used.